orchestrator/requirements.txt
azure-functions
azure-functions-durable>=1.2.9
azure-identity>=1.16.0
azure-mgmt-resource>=23.0.0
azure-mgmt-containerinstance>=10.1.0
azure-mgmt-containerregistry>=10.3.0
azure-mgmt-storage>=21.0.0
azure-mgmt-authorization>=4.0.0
azure-storage-blob>=12.19.0
azure-kusto-data>=4.4.0
requests>=2.31.0
pydantic>=2.6.0
